reorganize all code into multiple files for better organization (fixes #15)
